Modifies to run actions on any push operation; comments out pytest
I decided to change this to any push action. It should still be well within the free tier. If the repo gets super busy we can modify it. :)

I am commenting out pytest as part of the workflow for now since the tests seem broken. I will submit a separate PR to fix and re-incorporate pytest.