Moves mnemonic to .secrets file, updates documentation, updates tests
This is my solution to not have to worry about committing secrets, specifically  to git. We can create a  file, add that to , and update compose to use that as an env file.

I also updated the tests to load the variable from  instead of the config file and I believe everything is happy.